chore: moved feedbackform into shared component in footer that can be submitted on any page
